#11: Transfer validations, Exception handling and unit tests
(cherry picked from commit 3a0c64fba66a612c43779a29fb2f4412f2d6e6a3)

+++ src/main/java/com/hardworkers/moneyspent/Transfer/Transfer.java
@@ -6,7 +6,7 @@ import lombok.*;
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Set;
 
@@ -36,7 +36,7 @@ public class Transfer {
     private LocalDateTime dateTime = LocalDateTime.now();
 
     @NotNull
-    private BigInteger count;
+    private BigDecimal count;
 
     @EqualsAndHashCode.Exclude
     @ManyToMany
